ui: fix to prevent keyboard from appearing too soon on mobile web on iOS
- When the dialog opens, if the input is focused and the keyboard opens
too soon, then the dialog does not shift up into view, and also the
arrow to focus the next input is not active either.
We'll avoid this bug by preventing autofocus to prevent keyboard auto-
appearing altogether. It's not ideal (less efficient) to require another
press from the user like this, but at least it will eliminate the ugly
and potentially annoying buggy experience of having keyboard laying on
top of form field which we want to remain visible.